ml_predictor/data_collection/requirements.txt
pandas==2.2.3
numpy==1.26.4
xgboost==3.1.1
scikit-learn==1.7.2
matplotlib==3.10.1
black==24.10.0
pyyaml
seaborn